requirements.txt
# Self-Healing Software System v2.0 - Python Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Template engine & form parsing (for HTML dashboard)
jinja2>=3.1.0
python-multipart>=0.0.6

# WebSocket support
websockets>=12.0

# HTTP client for AI API calls and error replication
aiohttp>=3.9.0
httpx>=0.27.0

# Database
asyncpg>=0.29.0  # PostgreSQL async driver
redis>=5.0.0      # Redis with async support

# Password hashing
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0  # JWT tokens

# Async file operations
aiofiles>=24.1.0

# Environment variable management
python-dotenv>=1.0.0

# Data validation and serialization
pydantic>=2.5.0

# Rich console output for logging
rich>=13.7.0

# Type hints support
typing-extensions>=4.9.0

# AST parsing - tree-sitter with language grammars
tree-sitter>=0.21.0
tree-sitter-languages>=1.10.0  # Pre-built grammars for 30+ languages
# Individual language packages (optional, tree-sitter-languages covers most)
tree-sitter-python>=0.21.0
tree-sitter-javascript>=0.21.0

# JSON schema validation
jsonschema>=4.21.0

# Diff generation for code changes
diff-match-patch>=20230430

# Test framework
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Code formatting (development)
black>=24.1.0
isort>=5.13.0

# Microsoft AutoGen multi-agent AI framework
autogen-agentchat>=0.7.0
autogen-ext[openai]>=0.7.0

# Static type checking (development)
mypy>=1.8.0
